renamed signals-config to signal-config
