+++ ui/app/util/signal-config.ts
